Add Texture class and integrate into renderer
Introduce a Texture class (include/texture.h, src/texture.cpp) to encapsulate image loading, binding and cleanup. Update project.kmake to include src/texture.cpp and set warning level. Replace manual stb_image texture handling in src/main.cpp with Texture::load/bind/destroy, switch timing vars to double, add multiple cube positions and per-cube model transforms, and remove raw stb_image usage. Also: change Camera::processKeyboard to use double deltaTime (header/cpp), cast index count to GLsizei in mesh draw, and minor cleanup/includes.

+++ src/camera.cpp
@@ -49,7 +49,7 @@ namespace Cthulhu::Scene
         return fov;
     }
 
-    void Camera::processKeyboard(GLFWwindow *window, float deltaTime)
+    void Camera::processKeyboard(GLFWwindow *window, double deltaTime)
     {
         float currentSpeed = speed * deltaTime;
         if (glfwGetKey(window, GLFW_KEY_ESCAPE) == GLFW_PRESS)
